pyproject.toml
[project]
name = "jarvis-lite"
version = "0.2.0"
description = "GTD task management via Claude Code Telegram channel"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["mcp_server.py"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "psycopg2-binary>=2.9",
]
